// types.ts
export interface OfficeLocation {
  id: string;
  name: string;
  lat: number;
  lng: number;
  radius: number;
  qrCodeString: string;
}